% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glvmfit.R
\docType{package}
\name{glvmfit}
\alias{glvmfit}
\title{glvmfit: Methods to Assess Generalized Latent Variable Model Fit}
\description{
Provides residual global fit indices for generalized latent variable models.
}
