% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.srmr.R
\name{print.srmr}
\alias{print.srmr}
\title{Print}
\usage{
\method{print}{srmr}(object, comp = c("Total", "Correlation", "Variance", "Mean"))
}
\arguments{
\item{object}{R object of class srmr}

\item{comp}{components to print}
}
\description{
Prints output of SRMR class
}
\examples{
Sigma <- matrix(c(1.022, .550,  .622, .550, .928, .783, .622, .783, 1.150), 
                    nrow = 3)
S <- matrix(c(.770, .545, .515, .545, 1.003, .890, .515, .890, 1.211), 
            nrow = 3)
ybar <- c(2.516, 4.041, 5.021)
mu <- c(2.825, 3.877, 4.929)

print(srmr(S = S,  Sigma = Sigma, ybar = ybar, mu = mu))

}
