% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ResidualFitIndex.R
\docType{class}
\name{ResidualFitIndex-class}
\alias{ResidualFitIndex-class}
\title{An S4 class to represent a residual fit indices.}
\description{
An S4 class to represent a residual fit indices.
}
\section{Slots}{

\describe{
\item{\code{type}}{A length-one numeric vector}

\item{\code{resid}}{A length-one numeric vector}

\item{\code{ssr}}{A length-one numeric vector}

\item{\code{size}}{A length-one numeric vector}

\item{\code{index}}{}
}}

