% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ResidualFitIndex.R
\docType{class}
\name{ResidualFitIndices-class}
\alias{ResidualFitIndices-class}
\title{An S4 class to represent the set of residual fit indicies}
\description{
An S4 class to represent the set of residual fit indicies
}
\section{Slots}{

\describe{
\item{\code{RMR}}{}

\item{\code{SRMR}}{}

\item{\code{CRMR}}{}
}}

