% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ResidualFitIndices.R
\docType{class}
\name{ResidualFitIndices-class}
\alias{ResidualFitIndices-class}
\alias{show}
\alias{show,ResidualFitIndices-method}
\alias{details}
\alias{details,ResidualFitIndices-method}
\title{An S4 class to represent the set of residual fit indicies}
\usage{
show(object)

\S4method{show}{ResidualFitIndices}(object)

details(object, comp)

\S4method{details}{ResidualFitIndices}(object, comp = c("Total", "Covariance", "Variance", "Mean", "Total"))
}
\arguments{
\item{object}{R object of type \code{ResidualFitIndices}.}

\item{comp}{Character indicating the components to include.}
}
\description{
An S4 class to represent the set of residual fit indicies
}
\section{Slots}{

\describe{
\item{\code{sampleMoments}}{}

\item{\code{impliedMoments}}{}

\item{\code{RMR}}{}

\item{\code{SRMR}}{}

\item{\code{CRMR}}{}
}}

\note{
\code{comp} can be "Total" for overall fit indices, "Cov" for
covariance elements (off diagonals), "Var" for variance components (diagonal), and "Mean"
means.
}
